Refactored validation outside User entity file
