+++ server/src/entities/User.ts
@@ -1,5 +1,4 @@
 import { Entity, Column, PrimaryGeneratedColumn } from 'typeorm';
-import * as Joi from 'joi';
 
 @Entity({ name: 'users' })
 export class User {
@@ -18,13 +17,3 @@ export class User {
     @Column({ default: false })
     admin: boolean;
 }
-
-export const validate = (user: any): Joi.ValidationResult<any> => {
-    // TODO: Temporarily serves both as login and register validator
-    // in the future register should have separate validator
-    const schema = {
-        email: Joi.string().min(5).max(100).required().email(),
-        password: Joi.string().min(5).max(255).required()
-    };
-    return Joi.validate(user, schema);
-};
